+++ tests/testthat/test_images.R
@@ -12,17 +12,17 @@ test_that("We can list images", {
 test_that("We can get an image from a family", {
   skip_on_cran()
   
-  image <- gce_get_image_family(image_project = "google-containers", family = "gci-stable")
+  image <- gce_get_image_family(image_project = "google-containers", family = "cos-stable")
     
   expect_equal(image$kind, "compute#image")
-  expect_equal(image$family, "gci-stable")  
+  expect_equal(image$family, "cos-stable")  
   
 })
 
 test_that("We can get a specific image", {
   skip_on_cran()
   
-  image <- gce_get_image(image_project = "google-containers", image = "gci-stable-53-8530-85-0")
+  image <- gce_get_image(image_project = "google-containers", image = "cos-stable-53-8530-85-0")
   
   expect_equal(image$kind, "compute#image")
 })