+++ NEWS.md
@@ -7,6 +7,7 @@
 - Let Rstudio users be added with staff rights so they can install packages etc.
 - Add ability to specify disk size when creating a VM (#38) - thanks @jburos
 - Add firewall functions (#34)
+- Add a builder template to build docker images on a dedicated VM (#32)
 
 # googleComputeEngineR 0.1.0